More specs (IO-capture, Rainbow fixes)
For some reason, Rainbow doesn't monkey-patch String anymore (if it ever did),
so those specs were failing. This commit also adds a capture_io helper to the
spec support, and tests the output of a couple of lines of code, increasing
the coverage.

Incidentally, this commit was pre-approved with a rake ghundle.